requirements.txt
redshift-connector
pandas
numpy
python-dotenv
